Allow to config NGINX's settings at runtime
The updated nginx-run script will check for an environment variable
and use it to set NGINX's config file before starting it.

This will allow users to set parameters of the server in runtime
rather than image-build time.

We're supporting a single setting for now (`client_max_body_size`),
but this can easily extended to look for every environment variable
prefixed by `NGINX_RAILS__` and write all of them to the config file.